reorganized to fix tensorflow
