+++ nlp_primitives/conftest.py
@@ -2,7 +2,9 @@ import sys
 
 import pytest
 
-from nlp_primitives.universal_sentence_encoder import UniversalSentenceEncoder
+from nlp_primitives.tensorflow.universal_sentence_encoder import (
+    UniversalSentenceEncoder
+)
 
 
 @pytest.fixture(scope="session")
@@ -13,11 +15,6 @@ def universal_sentence_encoder(pytestconfig):
         return None
 
 
-@pytest.fixture(autouse=True)
-def add_primitives(doctest_namespace, universal_sentence_encoder):
-    doctest_namespace['universal_sentence_encoder'] = universal_sentence_encoder
-
-
 def pytest_addoption(parser):
     parser.addoption(
         "--notensorflow",
@@ -25,20 +22,6 @@ def pytest_addoption(parser):
         default=False,
         help="If true, tests will assume tensorflow is not installed",
     )
-
-
-def pytest_configure(config):
-    config.addinivalue_line(
-        "markers", "notensorflow: mark test to be skipped if tensorflow is not installed"
-    )
-
-
-def pytest_collection_modifyitems(config, items):
-    if config.getoption("--notensorflow"):
-        skip_notensorflow = pytest.mark.skip(reason="no tensorflow installed")
-        for item in items:
-            if "notensorflow" in item.keywords:
-                item.add_marker(skip_notensorflow)
 
 
 def pytest_ignore_collect(path, config):
