+++ nlp_primitives/tensorflow/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from .universal_sentence_encoder import Un
+
+
+@pytest.fixture(autouse=True)
+def add_primitives(doctest_namespace, universal_sentence_encoder):
+    doctest_namespace['universal_sentence_encoder'] = universal_sentence_encoder
