+++ nlp_primitives/tensorflow/elmo.py
